Cleanup: Move class ChatBot to own file.
